benchmark switch as widget
