+++ src/humbug/widgets/__init__.py
@@ -1,0 +1,1 @@
+"""Reusable Humbug widgets."""
