+++ src/humbug/widgets/benchmark_switch.py
@@ -1,0 +1,99 @@
+"""Benchmark switch widget."""
+
+from PySide6.QtCore import QEvent, QPoint, QSize, Qt
+from PySide6.QtGui import QColor, QFont, QMouseEvent, QPainter
+from PySide6.QtWidgets import QCheckBox, QWidget
+
+from humbug.color_role import ColorRole
+from humbug.style_manager import StyleManager
+
+
+class BenchmarkSwitch(QCheckBox):
+    """Compact switch control for benchmark mode."""
+
+    def __init__(self, parent: QWidget | None = None) -> None:
+        """Initialize the switch."""
+        super().__init__(parent)
+        self._track_on_color = QColor("#2050c0")
+        self._track_off_color = QColor("#303030")
+        self._track_border_color = QColor("#606060")
+        self._knob_color = QColor("#ffffff")
+        self._text_on_color = QColor("#ffffff")
+        self._text_off_color = QColor("#d0d0d0")
+        self._knob_inset = 3
+        self.setCursor(Qt.CursorShape.PointingHandCursor)
+        self.setFocusPolicy(Qt.FocusPolicy.StrongFocus)
+        self.setMinimumSize(44, 20)
+        self.toggled.connect(lambda _checked: self.update())
+
+    def apply_style(self, style_manager: StyleManager) -> None:
+        """Apply design-system colors and metrics."""
+        self._track_on_color = QColor(style_manager.get_color_str(ColorRole.SWITCH_BACKGROUND_ON))
+        self._track_off_color = QColor(style_manager.get_color_str(ColorRole.SWITCH_BACKGROUND_OFF))
+        self._track_border_color = QColor(style_manager.get_color_str(ColorRole.SWITCH_BORDER))
+        self._knob_color = QColor(style_manager.get_color_str(ColorRole.SWITCH_KNOB))
+        self._text_on_color = QColor(style_manager.get_color_str(ColorRole.SWITCH_TEXT_ON))
+        self._text_off_color = QColor(style_manager.get_color_str(ColorRole.SWITCH_TEXT_OFF))
+        self._knob_inset = style_manager.switch_knob_inset()
+        self.setFixedSize(style_manager.switch_width(), style_manager.switch_height())
+        self.update()
+
+    def sizeHint(self) -> QSize:
+        """Return the preferred switch size."""
+        return self.minimumSize()
+
+    def hitButton(self, pos: QPoint) -> bool:
+        """Make the whole pill clickable."""
+        return self.rect().contains(pos)
+
+    def mouseReleaseEvent(self, event: QMouseEvent) -> None:
+        """Toggle from anywhere inside the switch."""
+        if event.button() == Qt.MouseButton.LeftButton and self.rect().contains(event.position().toPoint()):
+            self.setChecked(not self.isChecked())
+            event.accept()
+            return
+
+        super().mouseReleaseEvent(event)
+
+    def paintEvent(self, event: QEvent) -> None:
+        """Paint the switch as a rounded pill."""
+        del event
+
+        painter = QPainter(self)
+        painter.setRenderHint(QPainter.RenderHint.Antialiasing)
+
+        track_rect = self.rect().adjusted(1, 1, -1, -1)
+        radius = track_rect.height() / 2
+        track_color = self._track_on_color if self.isChecked() else self._track_off_color
+        text_color = self._text_on_color if self.isChecked() else self._text_off_color
+
+        painter.setPen(self._track_border_color)
+        painter.setBrush(track_color)
+        painter.drawRoundedRect(track_rect, radius, radius)
+
+        knob_size = track_rect.height() - (self._knob_inset * 2)
+        knob_y = track_rect.y() + self._knob_inset
+        if self.isChecked():
+            knob_x = track_rect.right() - knob_size - self._knob_inset
+
+        else:
+            knob_x = track_rect.x() + self._knob_inset
+
+        painter.setPen(Qt.PenStyle.NoPen)
+        painter.setBrush(self._knob_color)
+        painter.drawEllipse(knob_x, knob_y, knob_size, knob_size)
+
+        label = "ON" if self.isChecked() else "OFF"
+        text_rect = track_rect.adjusted(6, 0, -6, 0)
+        if self.isChecked():
+            text_rect.setRight(knob_x - 2)
+
+        else:
+            text_rect.setLeft(knob_x + knob_size + 2)
+
+        font = QFont(painter.font())
+        font.setBold(True)
+        font.setPointSizeF(max(6.0, font.pointSizeF() * 0.68))
+        painter.setFont(font)
+        painter.setPen(text_color)
+        painter.drawText(text_rect, Qt.AlignmentFlag.AlignCenter, label)
